+++ examples/data-sources/iosxe_vlan_group/data-source.tf
@@ -1,3 +1,3 @@
 data "iosxe_vlan_group" "example" {
-  name = "group"
+  name = "GROUP1"
 }