+++ examples/resources/iosxe_vlan_filter/resource.tf
@@ -1,4 +1,4 @@
 resource "iosxe_vlan_filter" "example" {
-  word       = "f1"
+  word       = "VAM1"
   vlan_lists = [1]
 }