+++ examples/resources/iosxe_vlan_group/import.sh
@@ -1,1 +1,1 @@
-terraform import iosxe_vlan_group.example "Cisco-IOS-XE-native:native/vlan/Cisco-IOS-XE-vlan:group=group"
+terraform import iosxe_vlan_group.example "Cisco-IOS-XE-native:native/vlan/Cisco-IOS-XE-vlan:group=GROUP1"