+++ internal/provider/data_source_iosxe_vlan_group.go
@@ -52,7 +52,7 @@ func (d *VLANGroupDataSource) Metadata(_ context.Context, req datasource.Metadat
 func (d *VLANGroupDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: "This data source can read the VLAN group configuration.",
+		MarkdownDescription: "This data source can read the VLAN Group configuration.",
 
 		Attributes: map[string]schema.Attribute{
 			"device": schema.StringAttribute{